codeandmore

+++ src/compiler/types/generate-types.ts
@@ -21,7 +21,10 @@ export const generateTypes = async (
 ): Promise<void> => {
   if (!buildCtx.hasError) {
     await generateTypesOutput(config, compilerCtx, buildCtx, outputTarget);
-    await copyStencilCoreDts(config, compilerCtx);
+
+    if (!outputTarget.keepCoreRefs) {
+      await copyStencilCoreDts(config, compilerCtx);
+    }
   }
 };
 
@@ -51,15 +54,22 @@ const generateTypesOutput = async (
       const distPath = join(outputTarget.typesDir, relPath);
 
       const originalDtsContent = await compilerCtx.fs.readFile(srcDtsFile.absPath);
-      const distDtsContent = updateStencilTypesImports(outputTarget.typesDir, distPath, originalDtsContent);
 
-      await compilerCtx.fs.writeFile(distPath, distDtsContent);
+      if (outputTarget.keepCoreRefs) {
+        await compilerCtx.fs.writeFile(distPath, originalDtsContent);
+      } else {
+        const distDtsContent = updateStencilTypesImports(outputTarget.typesDir, distPath, originalDtsContent);
+        await compilerCtx.fs.writeFile(distPath, distDtsContent);
+      }
+
       distDtsFilePath = distPath;
     })
   );
 
-  const distPath = outputTarget.typesDir;
-  await generateAppTypes(config, compilerCtx, buildCtx, distPath);
+  if (!outputTarget.keepCoreRefs) {
+    const distPath = outputTarget.typesDir;
+    await generateAppTypes(config, compilerCtx, buildCtx, distPath);
+  }
 
   if (distDtsFilePath) {
     await generateCustomElementsTypes(config, compilerCtx, buildCtx, distDtsFilePath);
